refactor to configure all logging in logger.py
